@@ -1,5 +1,5 @@
 module github.com/ebarkie/goblusb
 
-go 1.15
+go 1.16
 
 require github.com/google/gousb v2.1.0+incompatible
